refactor(runbotics-desktop): :truck: Moved types from services into external files
