+++ runbotics/runbotics-desktop/src/core/auth/auth.types.ts
@@ -1,0 +1,6 @@
+export type TokenData = {
+    token: string,
+    expiresAt: number,
+    tenantId: string,
+    installationId: string
+}
