+++ runbotics/runbotics-desktop/src/core/websocket/auth/webSocketAuth.types.ts
@@ -1,0 +1,7 @@
+export type WebSocketAuthData = {
+    token: string
+    installationId: string
+    system: string
+    collection: string
+    version: string
+}
